refactor: Externalize data to config.jl and fix run.jl
This commit refactors the simulation data by moving all hardcoded constants and optimization tables from  into a new, dedicated  file. This improves code organization and maintainability.

Key changes include:
- : A new file containing , , , and all  tables, wrapped in a  module.
- :
    - Removed all hardcoded data tables.
    - Added  and  to import the data.
    - Corrected the  typo that caused the .
    - Updated  to correctly select optimization data based on  from .
