+++ config.jl
@@ -1,0 +1,141 @@
+# ===================================================================
+# CONFIGURATION DATA FOR PSA SIMULATOR
+# ===================================================================
+# This file contains all material properties and optimization 
+# variables from the reference paper.
+# ===================================================================
+
+module PSAConfigData
+
+export ISOTHERM_PARAMETERS, SIMULATION_PARAMETERS, MATERIALS_LIST, 
+       OPT_VARS_PURITY, OPT_VARS_RECOVERY, OPT_VARS_PRODUCTIVITY, OPT_VARS_ENERGY
+
+# ===================================================================
+# MATERIAL PROPERTIES
+# ===================================================================
+
+# Isotherm parameters from Params.mat
+const ISOTHERM_PARAMETERS = [
+    7.022572e+00 6.326170e+00 2.647463e-12 2.269914e-11 -3.375370e+04 -3.641874e+04 7.022572e+00 0.000000e+00 8.228112e-09 0.000000e+00 -1.200000e+04 0.000000e+00 0.000000e+00;
+    9.999929e+00 8.838786e+00 7.416992e-11 1.296440e-10 -2.493171e+04 -2.493171e+04 1.000000e+01 0.000000e+00 2.368616e-08 0.000000e+00 -5.800000e+03 0.000000e+00 0.000000e+00;
+    1.000000e+00 1.892000e+01 1.395845e-09 5.911813e-12 -3.300000e+04 -3.300000e+04 3.200000e+00 0.000000e+00 1.040135e-08 0.000000e+00 -1.200000e+04 0.000000e+00 0.000000e+00;
+    6.800000e+00 9.900000e+00 2.440000e-11 1.390000e-10 -4.200000e+04 -2.400000e+04 1.400000e+01 0.000000e+00 4.960000e-10 0.000000e+00 -1.800000e+04 0.000000e+00 0.000000e+00;
+    4.800000e+01 0.000000e+00 8.060000e-10 0.000000e+00 -1.400000e+04 -1.400000e+04 4.800000e+01 0.000000e+00 1.110000e-09 0.000000e+00 -1.000000e+04 0.000000e+00 0.000000e+00;
+    6.212000e+00 7.150000e+00 4.377195e-11 4.608793e-13 -3.785300e+04 -3.785300e+04 1.190000e+01 0.000000e+00 1.383811e-10 0.000000e+00 -1.943460e+04 0.000000e+00 0.000000e+00;
+    2.855382e+01 0.000000e+00 5.938529e-11 0.000000e+00 -2.500000e+04 -2.500000e+04 2.855382e+01 0.000000e+00 1.220402e-09 0.000000e+00 -1.200000e+04 0.000000e+00 0.000000e+00;
+    1.535652e+00 2.674846e+00 5.853452e-13 4.523258e-12 -3.250000e+04 -3.250000e+04 1.535652e+00 0.000000e+00 5.010345e-11 0.000000e+00 -1.800000e+04 0.000000e+00 0.000000e+00;
+    6.816578e+00 0.000000e+00 8.436744e-11 0.000000e+00 -3.200000e+04 -3.200000e+04 3.890598e+00 0.000000e+00 3.827598e-09 0.000000e+00 -1.200000e+04 0.000000e+00 0.000000e+00;
+    2.727278e+00 2.495993e-01 2.891278e-12 2.007587e-21 -5.209388e+04 -9.023652e+04 2.727278e+00 0.000000e+00 6.520000e-09 0.000000e+00 -1.200000e+04 0.000000e+00 0.000000e+00;
+    4.540935e+00 0.000000e+00 9.931281e-13 0.000000e+00 -4.200000e+04 -4.200000e+04 2.181907e+00 0.000000e+00 1.538952e-10 0.000000e+00 -1.900000e+04 0.000000e+00 0.000000e+00;
+    5.000000e+00 3.000000e+00 9.460000e-11 6.150000e-16 -3.300000e+04 -4.800000e+04 1.270000e+01 0.000000e+00 4.290000e-10 0.000000e+00 -1.230000e+04 0.000000e+00 0.000000e+00;
+    7.265438e+00 0.000000e+00 1.215474e-09 0.000000e+00 -1.612044e+04 -1.612044e+04 7.265438e+00 0.000000e+00 2.345892e-10 0.000000e+00 -7.020028e+03 0.000000e+00 0.000000e+00;
+    7.000000e+00 5.100000e+00 3.207140e-10 2.148947e-11 -2.750000e+04 -2.750000e+04 6.000000e+00 0.000000e+00 5.189774e-09 0.000000e+00 -1.000000e+04 0.000000e+00 0.000000e+00;
+    6.570000e+00 3.130000e+00 1.440000e-07 9.410000e-07 -2.924300e+04 -3.080000e+04 9.606000e+00 0.000000e+00 7.350000e-06 0.000000e+00 -1.290000e+04 0.000000e+00 1.000000e+00;
+    3.090000e+00 2.540000e+00 8.650000e-07 2.630000e-08 -3.664121e+04 -3.569066e+04 5.840000e+00 0.000000e+00 2.500000e-06 0.000000e+00 -1.580000e+04 0.000000e+00 1.000000e+00
+]
+
+# Simulation parameters from Params.mat
+const SIMULATION_PARAMETERS = [
+    1.169000e+03 -3.450000e+04 -1.200000e+04;
+    7.890000e+02 -2.490000e+04 -5.800000e+03;
+    7.820000e+02 -3.300000e+04 -1.200000e+04;
+    9.100000e+02 -3.130000e+04 -1.800000e+04;
+    4.270000e+02 -1.400000e+04 -1.000000e+04;
+    1.200000e+03 -3.785300e+04 -1.943460e+04;
+    6.260000e+02 -2.500000e+04 -1.200000e+04;
+    1.236000e+03 -3.250000e+04 -1.800000e+04;
+    8.500000e+02 -3.200000e+04 -1.200000e+04;
+    1.620000e+03 -5.530000e+04 -1.200000e+04;
+    1.770000e+03 -4.200000e+04 -1.900000e+04;
+    1.659000e+03 -3.300000e+04 -1.200000e+04;
+    9.500000e+02 -1.610000e+04 -7.000000e+03;
+    1.179000e+03 -2.750000e+04 -1.000000e+04;
+    1.230000e+03 -3.030000e+04 -1.290000e+04;
+    1.130000e+03 -3.600000e+04 -1.580000e+04
+]
+
+# ===================================================================
+# MATERIAL DEFINITIONS
+# ===================================================================
+
+const MATERIALS_LIST = [
+    (index=1, name="Co-MOF-74"), (index=2, name="Cu-BTTRi"), (index=3, name="Cu-TDPAT"),
+    (index=4, name="Mg-MOF-74"), (index=5, name="MOF-177"), (index=6, name="Ni-MOF-74"),
+    (index=7, name="NTU-105"), (index=8, name="Sc2BDC3"), (index=9, name="SIFSIX-2-Cu-i"),
+    (index=10, name="SIFSIX-3-Ni"), (index=11, name="Ti-MIL-91"), (index=12, name="USTA-16"),
+    (index=13, name="UiO-66(OH)2"), (index=14, name="ZIF-8"), (index=15, name="Zn-MOF-74"),
+    (index=16, name="Zeolite_13X")
+]
+
+# ===================================================================
+# OPTIMIZATION VARIABLES
+# ===================================================================
+# Format: [P_0, t_ads, alpha, beta, gamma, P_l] for each material
+
+# Table S4: Process optimization (max purity, 90% CO2 recovery)
+const OPT_VARS_PURITY = [
+    1.000000e+05 8.082500e+02 2.400000e-01 2.600000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 6.843800e+02 1.400000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.580000e+05 5.806300e+02 1.800000e-01 2.000000e-01 1.000000e+00 1.000000e+04;
+    2.050000e+05 5.110000e+02 1.400000e-01 2.700000e-01 1.000000e+00 1.000000e+04;
+    1.600000e+05 9.138000e+01 1.900000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.080000e+05 8.660400e+02 1.400000e-01 2.800000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 4.740900e+02 1.900000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 2.471600e+02 1.600000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 5.192200e+02 1.800000e-01 3.600000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 4.343700e+02 1.100000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.120000e+05 5.268500e+02 1.800000e-01 3.500000e-01 1.000000e+00 1.000000e+04;
+    1.420000e+05 8.283200e+02 1.100000e-01 3.500000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 6.921900e+02 2.000000e-01 3.600000e-01 1.000000e+00 1.000000e+04;
+    1.160000e+05 2.012900e+02 1.100000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 8.599000e+02 2.200000e-01 3.000000e-01 1.000000e+00 1.000000e+04;
+    1.640000e+05 4.865300e+02 1.000000e-01 3.000000e-01 1.000000e+00 1.000000e+04
+]
+
+# Table S5: Process optimization (max purity, 95% CO2 recovery)
+const OPT_VARS_RECOVERY = [
+    1.000000e+05 1.000000e+03 2.400000e-01 2.100000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 6.388400e+02 3.500000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.600000e+05 6.577200e+02 2.100000e-01 1.600000e-01 1.000000e+00 1.000000e+04;
+    1.420000e+05 4.373100e+02 2.200000e-01 3.400000e-01 1.000000e+00 1.000000e+04;
+    1.020000e+05 5.956000e+01 3.300000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.540000e+05 6.631200e+02 1.300000e-01 2.900000e-01 1.000000e+00 1.000000e+04;
+    1.010000e+05 3.969900e+02 2.200000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 2.249300e+02 1.400000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.040000e+05 7.347300e+02 2.100000e-01 2.200000e-01 1.000000e+00 1.000000e+04;
+    1.040000e+05 7.347300e+02 2.100000e-01 2.200000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 7.775100e+02 2.100000e-01 2.300000e-01 1.000000e+00 1.000000e+04;
+    3.540000e+05 6.933900e+02 8.000000e-02 2.500000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 8.349100e+02 2.000000e-01 2.800000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 1.681300e+02 1.800000e-01 1.000000e-01 1.000000e+00 1.000000e+04;
+    1.000000e+05 8.740800e+02 2.300000e-01 2.600000e-01 1.000000e+00 1.000000e+04;
+    1.400000e+05 5.921300e+02 1.600000e-01 2.400000e-01 1.000000e+00 1.000000e+04
+]
+
+# Table S10: Economic optimization (max productivity)
+# Note: Extra 7th column in productivity data (not used in simulation)
+const OPT_VARS_PRODUCTIVITY = [
+    4.991420e+05 6.363080e+01 1.152310e-01 9.098500e-01 1.000000e+00 1.000000e+04 2.520000e+00;
+    2.592710e+05 1.205380e+02 1.396240e-01 9.308590e-01 1.000000e+00 1.000000e+04 2.730000e+00;
+    5.771610e+05 9.622060e+01 1.128495e-01 7.509580e-01 1.000000e+00 1.000000e+04 3.360000e+00;
+    2.363850e+05 1.549430e+02 1.471220e-01 7.671240e-01 1.000000e+00 1.000000e+04 2.310000e+00;
+    3.704540e+05 9.813070e+01 9.545030e-02 1.032900e+00 1.000000e+00 1.000000e+04 3.420000e+00;
+    7.391500e+05 4.281040e+01 8.920810e-02 1.935910e+00 1.000000e+00 1.000000e+04 2.310000e+00;
+    2.325780e+05 2.145670e+02 1.184600e-01 8.304280e-01 1.000000e+00 1.000000e+04 3.420000e+00;
+    7.633360e+05 4.608960e+01 9.510270e-02 9.659200e-01 1.000000e+00 1.049790e+04 2.940000e+00
+]
+
+# Table S11: Economic optimization (min energy)
+# Note: Extra 7th column in energy data (not used in simulation)
+const OPT_VARS_ENERGY = [
+    1.000000e+05 2.159200e+02 1.167500e-01 4.323490e-01 1.000000e+00 1.430000e+04 1.800000e+00;
+    1.000000e+05 2.963020e+02 1.132530e-01 6.312100e-01 1.000000e+00 1.028490e+04 2.340000e+00;
+    1.000000e+05 6.884410e+02 7.806130e-02 3.048900e-01 1.000000e+00 1.092830e+04 2.520000e+00;
+    1.000000e+05 4.305300e+02 1.087420e-01 3.667430e-01 1.000000e+00 1.232570e+04 2.120000e+00;
+    1.000000e+05 4.098590e+02 8.116020e-02 3.597560e-01 1.000000e+00 1.180100e+04 2.160000e+00;
+    1.000000e+05 3.461020e+02 5.262560e-02 6.113690e-01 9.395290e-01 1.074040e+04 2.210000e+00;
+    1.000000e+05 5.137900e+02 1.054910e-01 4.380490e-01 1.000000e+00 1.151060e+04 2.310000e+00;
+    1.000000e+05 4.314640e+02 5.876720e-02 2.964420e-01 9.847090e-01 1.091640e+04 2.210000e+00
+]
+
+end # module PSAConfigData
